Moved compositing operations into plotting utilities
